fixed bugs

+++ headers/InteriorPoint.hpp
@@ -12,10 +12,10 @@ using namespace std;
 class InteriorPoint {
 public:
     static void start_interior_point(const Matrix& A, const Vector& b, const Vector& c,
-                                     const Vector& trial_solution, float alpha, float accuracy);
+                                     const Vector& trial_solution, double alpha, double accuracy);
 
 private:
-    static bool check_data(const Matrix &A, const Vector &b, const Vector &c, float alpha, float accuracy);
+    static bool check_data(const Matrix &A, const Vector &b, const Vector &c, double alpha, double accuracy);
 
     static void initialize_algorithm_data(const Matrix &A, const Vector &B, const Vector &C, Matrix &main_matrix,
                                           Vector &func_coefficients);
